+++ build-tools/packages/build-tools/src/common/fluidRepo.ts
@@ -54,6 +54,8 @@ export interface IFluidBuildConfig {
 	branchReleaseTypes?: {
 		[name: string]: VersionBumpType | PreviousVersionStyle;
 	};
+
+	typeValidation?: ITypeValidation;
 }
 
 /**
@@ -64,6 +66,54 @@ export interface PolicyConfig {
 	dependencies?: {
 		requireTilde?: string[];
 	};
+}
+
+/**
+ * Metadata about known-broken types.
+ */
+export interface BrokenCompatSettings {
+	backCompat?: false;
+	forwardCompat?: false;
+}
+
+/**
+ * A mapping of a type name to its {@link BrokenCompatSettings}.
+ */
+export type BrokenCompatTypes = Partial<Record<string, BrokenCompatSettings>>;
+
+export interface ITypeValidation {
+	/**
+	 * The version of the package. Should match the version field in package.json.
+	 */
+	version: string;
+
+	/**
+	 * An object containing types that are known to be broken.
+	 */
+	broken: BrokenCompatTypes;
+
+	/**
+	 * If true, disables type test preparation and generation for the package.
+	 */
+	disabled?: boolean;
+
+	/**
+	 * The previous version style that was used when the prepare phase was run. This value is cached so that
+	 * generation can work even on branches without the correct config.
+	 */
+	previousVersionStyle?: PreviousVersionStyle;
+
+	/**
+	 * The version range used as the "previous" version to compare against when generating type tests. This may be
+	 * an exact version or a range string.
+	 */
+	baselineRange?: string;
+
+	/**
+	 * The exact version used as the "previous" version to compare against when generating type tests. This should
+	 * always be an exact version.
+	 */
+	baselineVersion?: string;
 }
 
 /**
