+++ build-tools/packages/build-tools/src/common/npmPackage.ts
@@ -10,8 +10,7 @@ import * as path from "path";
 import sortPackageJson from "sort-package-json";
 
 import { options } from "../fluidBuild/options";
-import type { PreviousVersionStyle } from "../typeValidator/packageJson";
-import { IFluidBuildConfig } from "./fluidRepo";
+import { type IFluidBuildConfig, type ITypeValidation } from "./fluidRepo";
 import { defaultLogger } from "./logging";
 import { MonoRepo, MonoRepoKind, PackageManager } from "./monoRepo";
 import {
@@ -70,62 +69,15 @@ export interface PackageJson {
 	[key: string]: any;
 
 	/**
-	 * fluid-build config. Some properties only apply when set in the root or release group root package.json.
-	 */
-	fluidBuild?: IFluidBuildConfig;
-
-	/**
 	 * type compatibility test configuration. This only takes effect when set in the package.json of a package. Setting
 	 * it at the root of the repo or release group has no effect.
 	 */
-	typeValidation?: {
-		/**
-		 * The version of the package. Should match the version field in package.json.
-		 */
-		version: string;
-
-		/**
-		 * An object containing types that are known to be broken.
-		 */
-		broken: BrokenCompatTypes;
-
-		/**
-		 * If true, disables type test preparation and generation for the package.
-		 */
-		disabled?: boolean;
-
-		/**
-		 * The previous version style that was used when the prepare phase was run. This value is cached so that
-		 * generation can work even on branches without the correct config.
-		 */
-		previousVersionStyle?: PreviousVersionStyle;
-
-		/**
-		 * The version range used as the "previous" version to compare against when generating type tests. This may be
-		 * an exact version or a range string.
-		 */
-		baselineRange?: string;
-
-		/**
-		 * The exact version used as the "previous" version to compare against when generating type tests. This should
-		 * always be an exact version.
-		 */
-		baselineVersion?: string;
-	};
+	typeValidation?: ITypeValidation;
+	/**
+	 * fluid-build config. Some properties only apply when set in the root or release group root package.json.
+	 */
+	fluidBuild?: IFluidBuildConfig;
 }
-
-/**
- * Metadata about known-broken types.
- */
-export interface BrokenCompatSettings {
-	backCompat?: false;
-	forwardCompat?: false;
-}
-
-/**
- * A mapping of a type name to its {@link BrokenCompatSettings}.
- */
-export type BrokenCompatTypes = Partial<Record<string, BrokenCompatSettings>>;
 
 export class Package {
 	private static packageCount: number = 0;
